MEDIUM: ssl: stop on private key errors
Previously, when the key was in the PEM file, and an error was detected
on the key, it was ignored and the error could have been reported later.

This patch fixes the issue by returning the error just after the loading
attempt.

Not to be backported, could probably break things for people who
separated their keys but have problems in their PEM.