MEDIUM: logs: implement tag aliases
Tag aliases are a new way to wrote log-format tags, it allows maintain a
list of tags that are aliased to sample fetches.

The long-term goal is to remove the original tags implementation so
everything is available as a log-format.

In order to register tags aliases you just need to register an array of
tags / sample fetches this way:

	/* tags that are aliases for sample timers */
	static struct tag2lf_list tag2lf_timers = {ILH, {
	       { "Ta",   "txn.timer.total"      },
	       { "Tr",   "req.timer.hdr"        },
	       { NULL,   NULL                      },
	}};

	INITCALL1(STG_REGISTER, log_register_tag2lf, &tag2lf_timers);